+++ Engine/Systems/Camera/CameraSystem.cpp
@@ -103,8 +103,8 @@ void ::Camera::CameraSystem::onWindowSizeChanged(
                            ->getSubSystem<EngineStore>()
                            ->getRoot()
                            .getComponentsOfType<Component::Camera>()) {
-    camera->viewport.width = width;
-    camera->viewport.height = height;
+    camera->viewport.width = static_cast<float>(width);
+    camera->viewport.height = static_cast<float>(height);
     camera->is_dirty = true;
   }
 }
